+++ apps/connect/source/ftrack_connect/application_launcher/discover_applications.py
@@ -1,3 +1,6 @@
+# :coding: utf-8
+# :copyright: Copyright (c) 2024 ftrack
+
 import sys
 import os
 import platform